Fix : Error [ - Setup SFML 32 bit ]
Edit : Keyboard.h[
- Not init keyboard_uniq
]

+++ SFMLTemplateGame1/Game.h
@@ -18,7 +18,7 @@ class Game {
 		void handleKeyboardUniq(const sf::Keyboard::Key key, bool state);
 		void handleMouse(const sf::Mouse::Button btn, bool state);
 		void handleMouseUniq(const sf::Mouse::Button btn, bool state);
-		void handleMouseScrolled(const sf::Mouse::Wheel wheel , float deltascr);
+		void handleMouseScrolled(const sf::Mouse::Wheel wheel , int deltascr);
 		void processBall();
 		void update();
 		void process();
